third_party/WebKit/LayoutTests/editing/execCommand/insert-paragraph-into-table.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(()=> assert_selection(
  [
  "<div contenteditable>",
    "<table>",
      "<tbody>",
        "<td style='display:inline-block'>",
          "<ruby><rt>a|</rt></ruby>",
        "</td>",
      "</tbody>",
    "</table>^",
  "</div>"],
  'InsertParagraph',
  [
  "<div contenteditable>",
    "<table>",
      "<tbody>",
        "<tr><td style=\"display:inline-block\">",
          "<ruby><rt>a</rt></ruby>",
        "</td></tr>",
        "<tr><span style=\"font-size: 8px;\">|<br></span></tr>",
      "</tbody>",
    "</table>",
  "</div>"
  ]), "");
</script>
